==> 2-data-cleansing.ipynb <==
# Chapter 2. Data cleansing

Now that we got the needed data, it's time to look at it to figure out whether it is ready to be used in our algorithms (spoiler: it is not). 

## Imports

At first we have some pretty standard modules here: `pandas`, `numpy` and `utils`, our file with custom functions. 

We're also going to keep our data in a .json file rather than .csv, and we'll need `json` module to perform the corresponding conversion. 

Built-in pandas functions for regular expressions will be enough for us almost everywhere, but in some cases we'll still need `re` module.

import pandas as pd
import numpy as np
import utils
import json
import re

## Captions transformation

The first step in our work today is to read the data from our `videos.csv` file and look at what we have there.

df = pd.read_csv("./data/videos.csv", index_col='Unnamed: 0')
df.drop(columns=['id'], inplace=True)

df.head()

df.shape

The first thing I would like to cope with is subtitles, as they have a major importance for the aims of the project. Let's take a look at what we've got in the previous chapter.

captions = df['captions']
captions.iloc[-1]

Expectedly, the captions now are a complete mess. It's not just simple pretty text data, it's made up of scary numbered lines with time codes. That's something we have to deal with, and slicing seems to be a really helpful and simple solution.

def transform_captions(caption):
    caption = caption.split(sep="\n")
    caption = caption[::-4]
    caption.reverse()
    caption = ' '.join(caption)
    return caption

captions = captions.apply(transform_captions)
print(captions.iloc[-1])

Well, now it looks like a normal edible text rather than that inconvenient thing with timestamps. We can replace our `captions` column in the dataframe with what we have now.

df['captions'] = captions

## Video titles and speaker names
The next step that I'd take is looking at the titles. Are they in an appropriate form? Of course they are not, that would be too simple.

df_temp = df.copy()

for i in range(10):
    print(df_temp.iloc[i, 0])

The problem with these titles is that they contain speakers' names along with the videos' "names". In order to have a clean and pretty dataset we're going to extract names into a separate column and remove them from our `title` column. That's where regular expressions come in handy.

vertical_bar_names = df_temp['title'].str.extract(r'\|\s*(.+)$')

Using the regular expression above we've extracted all symbols that were met after `|` symbol in our `title` column, since 10 most recent videos' titles are compiled in accordance with this scheme. Let's see if that's enough.

def count_extracted_names(speaker_series, df):
    print(f"Number of speaker names extracted: {speaker_series.notnull().sum().iloc[0]}/{df.shape[0]}")

for i in range(10):
    print(vertical_bar_names[vertical_bar_names[0].notnull()].iloc[i, 0])

count_extracted_names(vertical_bar_names, df_temp)

We have dealt with the most of video titles, but there are still many left. It means that not every video title corresponds to the form we have expected.  

df_temp['speaker_name'] = vertical_bar_names
df_temp = df_temp[df_temp['speaker_name'].isna()]

for i in range(10):
    print(df_temp[df_temp['speaker_name'].isna()].iloc[-i, 0])

The situation is getting worse: we will need a more complicated regex to cope with those names that go at the beginning of the line. But at least we see that they are usually followed by a colon.

Note the fact that we are going to apply next regex only to the rows from which we've failed to extract speaker names earlier. That is why we have created a temporary DataFrame `df_temp` and dropped the rows with non-null `speaker_name` values.

colon_names = df_temp['title'].str.extract(r'^((?:[A-Z]{1}[^:\(\)]+){1,}):.*$')

for i in range(10):
    print(colon_names[colon_names[0].notnull()].iloc[-i, 0])

count_extracted_names(colon_names, df_temp)

df_temp['speaker_name'] = colon_names
df_temp = df_temp[df_temp['speaker_name'].isna()]

for i in range(10):
    print(df_temp.iloc[i, 0])

Now that we are done with our so-called colon names, we can use an apostrophe after some other names as a break point. 

apostrophe_names = df_temp['title'].str.extract(r'^((?:[A-Z]{1}[-\w.]+\s?)+)\'s?.*$')

for i in range(10):
    print(apostrophe_names[apostrophe_names[0].notnull()].iloc[-i, 0])

count_extracted_names(apostrophe_names, df_temp)

df_temp['speaker_name'] = apostrophe_names
df_temp = df_temp[df_temp['speaker_name'].isna()]

for i in range(10, 20):
    print(df_temp.iloc[i, 0])

It's time to deal with the names that are not followed by anything at all. These may bring some mistakes, since we do not have a clear break point. The capitalization is our means to differ names from anything else. However, we will be unable to capture one-word names because literally every first word in titles is going to be capitalized.

no_breakpoint_names = df_temp['title'].str.extract(r'^((?:[A-Z]{1}[-a-z.]+\s?(?:and|with)?\s?){2,}).*$')

for i in range(10, 20):
    print(no_breakpoint_names[no_breakpoint_names[0].notnull()].iloc[-i, 0])

count_extracted_names(no_breakpoint_names, df_temp)

df_temp['speaker_name'] = no_breakpoint_names
df_temp = df_temp[df_temp['speaker_name'].isna()]

Look, we have only failed to extract 19 names out of 3589!  
We can try to deal with them manually or just drop. In any case, we now need to gather up the new data in our good old `df` dataframe.

for i in df_temp.index:
    print(i, df_temp.loc[i, 'title'])

names_list = [vertical_bar_names, colon_names, apostrophe_names, no_breakpoint_names]

for x in names_list:
    x.dropna(inplace=True)
    
sp = pd.concat(names_list)
df['speaker_name'] = sp

df.loc[252, 'speaker_name'] = 'Aparna Pallavi'
df.loc[2306, 'speaker_name'] = 'iO Tillett Wright'
df.loc[2558, 'speaker_name'] = 'Morley'
df.loc[3133, 'speaker_name'] = 'Charles + Ray Eames'

df = df[df['speaker_name'].notnull()]
df.reset_index(drop=True, inplace=True)
df.shape

Let's not forget that we must also remove names we've extracted from the title column along with some special characters like `|+: ` and space characters. 

for i in range(10):
    print(i, df.loc[i, 'title'])

unique_names = list(df.speaker_name.unique())
unique_names = [x for x in unique_names if str(x) != 'nan']

for name in unique_names:
    df['title'] = df['title'].str.replace(name, '', regex=False) 

df['title'] = df['title'].str.replace(r'(\||\+|:)', '', regex=True)
df['title'] = df['title'].str.replace(' ?\(.*\)', '', regex=True)

df['title'] = df['title'].str.replace(r'^\'s?\s?', ' ', regex=True)

df['title'] = df['title'].str.replace(r'\s{2,}', ' ', regex=True)
df['title'] = df['title'].str.replace(r'^\s', '', regex=True)
df['title'] = df['title'].str.replace(r'\s$', '', regex=True)

df['speaker_name'] = df['speaker_name'].str.replace(r'\|? ted ?talks?', '', flags=re.IGNORECASE, regex=True)
df['speaker_name'] = df['speaker_name'].str.replace(r'(?: with| and | \+)', ',', regex=True)
df['speaker_name'] = df['speaker_name'].str.replace(r'(?:\(|\))', '', regex=True)

for i in range(3133, 3143):
    print(i, df.loc[i, 'title'])

Now we are officially done with speaker names and titles. And it means that it's time to move on to the next issue.

## Transcribers and reviewers

That is a small task that requires a bit more work with subtitles. There's some information about those who transcribed and rewieved the subtitles right inside the `captions` column. It is pretty improbable that we will make a good use of it, but let's extract this data just to be nice (and to have a detailed and thorough dataset).  

I think that it's easier to open a file with raw captions once again so as to have end-of-line symbols that are pretty useful for us now.

raw = pd.read_csv("./data/temp/raw_captions.csv", index_col='Unnamed: 0')
raw.rename(columns = {'0': 'captions'}, inplace = True)

transcriber = raw['captions'].str.extract(r'Transcriber: ?(.+)(?:Reviewer:.*)\n', flags=re.IGNORECASE)
reviewer = raw['captions'].str.extract(r'Transcriber:.+Reviewer: ?(.+)\n', flags=re.IGNORECASE)

transcriber[transcriber[0].notnull()].shape, reviewer[reviewer[0].notnull()].shape

df['transcriber'] = transcriber
df['reviewer'] = reviewer

unique_transcribers = list(transcriber[0].unique())
unique_reviewers = list(reviewer[0].unique())
unique_transcribers.reverse(), unique_reviewers.reverse()
unique_transcribers.pop(), unique_reviewers.pop()

df['captions'] = df['captions'].str.replace('Transcriber:', '', flags=re.IGNORECASE)
df['captions'] = df['captions'].str.replace('Reviewer:', '', flags=re.IGNORECASE)
for name in unique_transcribers:
    df['captions'] = df['captions'].str.replace(name, '')
for name in unique_reviewers:
    df['captions'] = df['captions'].str.replace(name, '')

df.iloc[103:108, [0, 13, 14]]

## Working with descriptions

Now it's time for some tedious regular expression work with video descriptions. As you can see below, the descriptions contain many chunks of information that are either non-unique or just are not related to the topic of the video directly. 

for i in range(3):
    print(df.iloc[i, 1])
    print("\n--------------------------\n")

df_checkpoint = df.copy()

Here we should take a look at different variations of paragraphs that are useless for us and create some regexes in order to delete them from our dataset. 

df['description'] = df['description'].str.replace(r'Visit http://TED\.com[^\n\n]*', '', \
                                                  flags=re.IGNORECASE, regex=True)
df['description'] = df['description'].str.replace(r'TEDTalks is a daily video podcast [^\n\n]*', '', \
                                                  flags=re.IGNORECASE, regex=True)
df['description'] = df['description'].str.replace(r'[^\n\n]*TED Talks channel features [^\n\n]*', '', \
                                                  flags=re.IGNORECASE, regex=True)
df['description'] = df['description'].str.replace(r'note: comments [^\n\n]*', '', \
                                                  flags=re.IGNORECASE, regex=True)
df['description'] = df['description'].str.replace(r'[^\n\n]* CC button [^\n\n]*', '', regex=True)
df['description'] = df['description'].str.replace(r'[^\n\n]*Watch the full talk[^\n\n]*', '', regex=True)
df['description'] = df['description'].str.replace('http://www.ted.com', '', regex=False)
df['description'] = df['description'].str.replace(r'[^\n]*https?://[^\n]*', '', regex=True)
df['description'] = df['description'].str.replace(r'[^\n]*\d{1,2}:\d{2}[^\n]*', '', regex=True)
df['description'] = df['description'].str.replace(r'[^\n]*A clip from[^\n]*', '', regex=True)
df['description'] = df['description'].str.replace(r'[^\n]*Think you know how your body works?[^\n]*', '', regex=True)
df['description'] = df['description'].str.replace(r'Find closed captions [^\n\n]*', '', regex=True)
df['description'] = df['description'].str.replace(r'\(?NOTE [^\n\n]*', '', \
                                                  flags=re.IGNORECASE, regex=True)
df['description'] = df['description'].str.replace('Take action on climate change at http://countdown.ted.com.', '', \
                                                  flags=re.IGNORECASE, regex=False)
df['description'] = df['description'].str.replace('Check out more TED Talks:', '', \
                                                  regex=False, flags=re.IGNORECASE)
df['description'] = df['description'].str.replace('/index.php/talks/top10', '', \
                                                  regex=False, flags=re.IGNORECASE)
df['description'] = df['description'].str.replace('www.ted.com', '', \
                                                  regex=False, flags=re.IGNORECASE)

We would also like to perform some operations with repetitive spaces and new-line symbols, or those that go in the end or beginning.  
It will make our data clearer.

columns = ['speaker_name', 'transcriber', 'reviewer', 'captions', 'description']

for col in columns:
    df[col] = df[col].str.replace('\n{2,}', '\n', regex=True)
    df[col] = df[col].str.replace('\n+$', '', regex=True)
    df[col] = df[col].str.replace('^\n+', '', regex=True)
    df[col] = df[col].str.replace('\s{2,}', ' ', regex=True)
    df[col] = df[col].str.replace('^\s', '', regex=True)
    df[col] = df[col].str.replace('\s$', '', regex=True)

for i in range(3):
    print(df.iloc[i, 1])
    print("\n--------------------------\n")

Here are the results of our work! Now the descriptions look informative and tidy.

## Extracting date and time

We are also going to do something with our `published_at` column to make it easier to work with.

df.published_at[0], df.published_at[2442]

Fortunately, date and time of video publication are presented in a constant form, which means that regexes will save our time and sanity once again.

datetime_r = r"^(\d{4})-(\d{2})-(\d{2})T(\d{2}):(\d{2}):\d{2}Z$"
date_time = df['published_at'].str.extract(datetime_r)

date_time.rename(columns={0:'year', 1:'month', 2:'day', 3:'hour', 4:'minute'}, inplace=True)
df = pd.concat([df, date_time], axis=1)
df.drop(columns='published_at', inplace=True)

df.iloc[:3, [0, 14, 15, 16, 17, 18]]

This is it, really simple after everything we have done previously.

## Json conversion

The last thing to do today is going to be some tags work. But right before that we need to transform our Pandas DataFrame into a .json object so as to preserve the dictionary structure of the data.

df = df[['video_id', 'channel_id', 'title', 'speaker_name', 'year', 'month', 'day', 'hour', 'minute', 'description', \
         'thumbnail', 'captions', 'transcriber', 'reviewer', 'views', 'likes', 'dislikes', 'comments_count', 'tags']]

df.to_csv('./data/videos_cleansed.csv')
utils.upload_to_googledrive('videos_cleansed.csv')

data_json = df.to_json(orient="records")
data_json = json.loads(data_json)

pd.json_normalize(data_json, max_level=0).tail(2)

By the way, we are still able to normalize our json object back to a dataframe for better visibility (if we need).

## Tag labelling

Finally, it's time to make dictionaries out of our tags. These dictionaries will contain not only tags themselves, but their types as well. Tags that contain something about "ted talks" will be labeled as garbage as they don't actually tell us anything important about the video, and tags that contain names of the speakers will be labeled as "name" tags.

all_tags = []

for i in range(df.shape[0]):
    tags = data_json[i]['tags'].strip('\'][').split('\', \'')
    tags_list = []
    full_name = (data_json[i]['speaker_name']).split()
    
    for tag in tags:
        
        any_ted = re.findall(r'.*ted.*', tag, re.IGNORECASE)
        any_talk = re.findall(r'.*talks?.*', tag, re.IGNORECASE)
        
        tag_type = 'normal'

        if (len(any_ted) != 0) | (len(any_talk) != 0):
            tag_type = 'garbage'    
        else: 
            for name in full_name:
                any_name = re.findall(name, tag)
                if (len(any_name) != 0):
                    tag_type = 'name'                
        
        tag_dict = {'tag': tag, 'type': tag_type}
        tags_list.append(tag_dict)
        
    all_tags.append(tags_list)

for i in range(len(all_tags)):
    data_json[i]['tags'] = all_tags[i]

print(f"{data_json[3441]['speaker_name']}, '{data_json[3441]['title']}'")
data_json[3441]['tags']

As you can observe, the work is done perfectly: tags connected with TED Talks were labeled as "garbage", speakers were labeled as "name", and everything else was claimed to be "normal".

## Making a dictionary out of date and time

Since we now keep our data in a json object, that would be logical to transform anything we can into dictionaries. It seems pretty natural to keep everything we know about date and time in the same place.

df = pd.json_normalize(data_json, max_level=0)
date_list = []

for i in range(df.shape[0]):
    year = df.iloc[i, 4]
    month = df.iloc[i, 5]
    day = df.iloc[i, 6]
    hour = df.iloc[i, 7]
    minute = df.iloc[i, 8]
    
    one_date = {'year': year, 'month': month, 'day': day,
               'hour': hour, 'minute': minute}
    date_list.append(one_date)
    
df['published_at'] = date_list
df.drop(columns=['year', 'month', 'day', 'hour', 'minute'], inplace=True)
df = df[['video_id', 'channel_id', 'title', 'speaker_name', 'published_at', 'description', \
         'thumbnail', 'captions', 'transcriber', 'reviewer', 'views', 'likes', 'dislikes', 'comments_count', 'tags']]

data_json = df.to_json(orient="records")
data_json = json.loads(data_json)

data_json[3573]

with open('./data/videos_clean_tags.json', 'w') as file:
    json.dump(data_json, file)

utils.upload_to_googledrive('videos_clean_tags.json')

df.info()

Now that we are done with data cleansing, it is time to move on to other kinds of tasks.
