==> 4-tf-idf-tags.ipynb <==
# Chapter 4. TF-IDF tags generation

The idea of this notebook is to get tags using TF-IDF of the terms. 

What we're going to do is to count a TF-IDF value for each word in each document; top-5 words with the highest values will become tags (since they are the most valuable for the document).

import numpy as np
import pandas as pd
import json
import nltk
import string
import utils
from nltk.tokenize import TreebankWordTokenizer
from nltk.stem import WordNetLemmatizer
from nltk.corpus import wordnet as wn
from nltk import word_tokenize, pos_tag
from collections import defaultdict

from sklearn.feature_extraction.text import ENGLISH_STOP_WORDS as sklearn_stop_words

The following downloads may be needed:

nltk.download('stopwords')
nltk.download('punkt')
nltk.download('wordnet')
nltk.download('averaged_perceptron_tagger');

nltk_stop_words = nltk.corpus.stopwords.words('english')
stop_words = list(set().union(nltk_stop_words, sklearn_stop_words))

with open('./data/videos_clean_tags.json') as json_file:
    data = json.load(json_file)

tokenizer = TreebankWordTokenizer()
lemmatizer = WordNetLemmatizer()

punctuation = ['``', '--', '""', "''", "?!", "...", "–"]
punctuation.extend(string.punctuation)

## Tokenizing

Here we tokenize our corpus, remove stopwords and lower the first letter of each sentence. Stemming is not used since we want our tags to be actual words, not stems. Lemmatization is used so as not to get many similar tags that have the same semantics but different forms. N-grams are also not used here for now.

def text_tokenize(text, to_lemmatize=False):
    tokenized = []
    tokenized_text = nltk.tokenize.sent_tokenize(text)
    ignore = list(set().union(stop_words, punctuation))
    
    for s in tokenized_text:
        s = s[0].lower() + s[1:]
        tokenized_sent = tokenizer.tokenize(s)
        tokenized_sent = [w for w in tokenized_sent if w.lower() not in ignore]
        tokenized.append(tokenized_sent)
        
    tokenized = [item for sublist in tokenized for item in sublist]
    
    if to_lemmatize:
        tag_map = defaultdict(lambda : wn.NOUN)
        tag_map['J'] = wn.ADJ
        tag_map['V'] = wn.VERB
        tag_map['R'] = wn.ADV

        lemmatized = []
        for token, tag in pos_tag(tokenized):
            lemma = lemmatizer.lemmatize(token, tag_map[tag[0]])
            lemmatized.append(lemma)
        tokenized = lemmatized
        
    return tokenized

%%time

for i in range (0, len(data)):
    data[i]['captions_tokenized'] = text_tokenize(data[i]['captions'], to_lemmatize=True)

This is not the most effective implementation of normalized term frequencies, but it's pretty clear in its steps.

full_lexicon = {}

for document in data:
    doc_lexicon = {}
    
    for term in document['captions_tokenized']:
        if doc_lexicon.get(term) is None:
            doc_lexicon[term] = 1            
        else:
            doc_lexicon[term] = doc_lexicon[term] + 1
            
    norm_term_frequencies = {}
    doc_length = len(document)
    
    for term in doc_lexicon.keys():
        norm_term_frequencies[term] = doc_lexicon[term]/doc_length
        
        if full_lexicon.get(term) is None:
            full_lexicon[term] = 1            
        else:
            full_lexicon[term] = full_lexicon[term] + 1
        
    document['norm_term_frequencies'] = norm_term_frequencies       

Let's see what we can achieve with normalized term frequencies only. 

sorted_dict = sorted(data[1]['norm_term_frequencies'].items(), key=lambda x: x[1])
sorted_dict.reverse()
sorted_dict[:10]

Actually, the results really help us understand the topic of the text even without the IDF part.

num_of_docs = len(data)

for document in data:
    tf_idf = {}
    for term in document['norm_term_frequencies'].keys():
        idf = np.log(num_of_docs/full_lexicon[term])
        tf_idf[term] = document['norm_term_frequencies'][term] * idf
        
    tf_idf_sorted = sorted(tf_idf.items(), key=lambda x: x[1])
    tf_idf_sorted.reverse()
    document['tf_idf_predicted_tags'] = tf_idf_sorted[:5]
    document.pop('norm_term_frequencies')
    document.pop('captions_tokenized')

for document in data:
    normal_tags = [row['tag'] for row in document['tags'] if row['type'] == 'normal']
    document['normal_tags'] = normal_tags

n = 427
data[n]['tf_idf_predicted_tags'], data[n]['normal_tags']

However, now, with the use of TF-IDF, the generated tags are even better; though the actual tags are usually different and less concrete.

with open('./data/tfidf_predicted_tags.json', 'w') as file:
    json.dump(data, file)

utils.upload_to_googledrive('tfidf_predicted_tags.json')
