==> 5-SVR-with-TF-IDF.ipynb <==
# Chapter 4. Generating tags using SVR

Using TF-IDF and word2vec Google pre-trained model.


import numpy as np
import pandas as pd
import random
import json
import sklearn
import gensim # using v.8.3.1
import nltk 
import logging
import utils
from sklearn.metrics import mean_squared_error as mse
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestRegressor, GradientBoostingRegressor
from gensim.models import KeyedVectors
from nltk import sent_tokenize, word_tokenize
from nltk.corpus import stopwords
from sklearn.feature_extraction.text import ENGLISH_STOP_WORDS as sklearn_stop_words
from sklearn.svm import SVR
from sklearn import linear_model

The following downloads may be needed:

nltk.download('punkt')
nltk.download('stopwords');

nltk_stop_words = nltk.corpus.stopwords.words('english')
stop_words = list(set().union(nltk_stop_words, sklearn_stop_words))

with open(utils.DATA + 'tfidf_predicted_tags.json') as json_file:
    data = json.load(json_file)

corpus, target = [], []

for doc in data:
    if len(doc['normal_tags']) != 0:
        corpus.append(doc['captions'])
        target.append(doc['normal_tags'])

print(f"Corpus length: {len(corpus)}")

df = pd.DataFrame(zip(corpus, target), columns=['doc', 'tag'])

train_data, test_data = train_test_split(df, test_size=0.1, random_state=42)

len(test_data)

%%time

wv = KeyedVectors.load_word2vec_format(utils.DATA + "w2v_googlenews/GoogleNews-vectors-negative300.bin.gz", binary=True)
wv.init_sims(replace=True)

First of all, we need to tokenize our documents and get a word2vec vector representation for each word. After that we're going to find the mean of all word vectors in a document so as to create one vector that is going to represent the text. And, finally, we will get the most similiar value from word2vec's vocabulary and see if this naive representation makes any sense.

def w2v_tokenize_tags(tag_list):
    sent = ""
    for tag in tag_list:
        sent = " ".join([sent, tag])
    tokens = w2v_tokenize_text(sent)
    return tokens

def w2v_tokenize_text(text):
    tokens = []
    for sent in sent_tokenize(text, language='english'):
        for word in word_tokenize(sent, language='english'):
            if len(word) < 2:
                continue
            if word in stop_words:
                continue
            tokens.append(word)
    return tokens

def w2v_transform_word(word):
    try:
        vector = wv.syn0norm[wv.vocab[word].index]
    except Exception as exception:
        # logging.warning(f"{type(exception).__name__} occured for word '{word}'.")
        vector = np.zeros(wv.vector_size,)
    return vector

def word_averaging(wv, words):
    all_words, mean = set(), []
    
    for word in words:
        if isinstance(word, np.ndarray):
            mean.append(word)
        elif word in wv.vocab:
            mean.append(wv.syn0norm[wv.vocab[word].index])
            all_words.add(wv.vocab[word].index)

    if not mean:
        logging.warning(f"cannot compute similarity with no input {words}")
        return np.zeros(wv.vector_size,)

    mean = gensim.matutils.unitvec(np.array(mean).mean(axis=0)).astype(np.float32)
    return mean

def word_averaging_list(wv, doc_list):
    return np.vstack([word_averaging(wv, doc) for doc in doc_list])

test_tokenized = test_data.apply(lambda r: w2v_tokenize_text(r['doc']), axis=1).values
train_tokenized = train_data.apply(lambda r: w2v_tokenize_text(r['doc']), axis=1).values

y_test_tokenized = test_data.apply(lambda r: w2v_tokenize_tags(r['tag']), axis=1).values
y_train_tokenized = train_data.apply(lambda r: w2v_tokenize_tags(r['tag']), axis=1).values

%%time

X_train_word_average = word_averaging_list(wv, train_tokenized)
X_test_word_average = word_averaging_list(wv, test_tokenized)

X_train_word_average.shape

y_train_word_average = word_averaging_list(wv, y_train_tokenized)
y_test_word_average = word_averaging_list(wv, y_test_tokenized)

wv.most_similar(positive=[X_test_word_average[100]], restrict_vocab=100000, topn=10)

As we can see, the naive method isn't really helpful, since the text simlarity is close to words that are evidentally aren't important for understanding the topic. So now we're going to try something else.

## Tag prediction based on TF-IDF 

Here we're going to make an assumption that words in each text that were considered the most important ones by the TF-IDF algorithm represent our documents well enough. So we will convert top-1 words (according to TF-IDF) into w2v vectors and use them to predict one tag that we already have in our dataframe as a target value.

### Preparing data

df

tf_idf_data = []

for n in range(len(data)):
    if (len(data[n]['normal_tags']) != 0) & (len(data[n]['tf_idf_predicted_tags']) >= 5):
        row = {}
        for i in range(5):
            tokens = []
            line = data[n]['tf_idf_predicted_tags'][i][0]
            for word in word_tokenize(line, language='english'):
                line = line.replace(' ', '_')
            row[f'top_word_{i+1}'] = line
        line = data[n]['normal_tags'][0]
        row['target'] = line.replace(' ', '_')
        tf_idf_data.append(row)

print(f"Total length: {len(tf_idf_data)}")

tfidf_df = pd.DataFrame(tf_idf_data)

tfidf_df = tfidf_df[['top_word_1', 'target']]

tfidf_df.head()

vectors = []

for col_name in tfidf_df.columns:
    col = tfidf_df.apply(lambda r: w2v_transform_word(r[col_name]), axis=1).values
    vectors.append(col)

df_temp = pd.DataFrame({'top-word': vectors[0], 'target': vectors[1]})

df_temp.head()

data_flat = pd.DataFrame(df_temp['top-word'].to_list())
y_flat = pd.DataFrame(df_temp['target'].to_list())

y_short = y_flat[:][0]
data_flat['target'] = y_short

data_flat.head()

### Trying to predict one of the coordinates of the target vector

train_df, test_df = train_test_split(data_flat, test_size=0.1, random_state=42)

y_train = train_df['target']
X_train = train_df.drop(columns=['target'])
y_test = test_df['target']
X_test = test_df.drop(columns=['target'])

%%time

svr = SVR(kernel='poly', degree=4, epsilon=0.05, C=3).fit(X_train, y_train)

y_pred_train = svr.predict(X_train)
y_pred_test = svr.predict(X_test)

print(f"MSE on train data:\t{mse(y_train, y_pred_train)}")
print(f"MSE on test data:\t{mse(y_test, y_pred_test)}")

%%time

linreg = linear_model.LinearRegression(n_jobs=1).fit(X_train, y_train)

y_pred_train = linreg.predict(X_train)
y_pred_test = linreg.predict(X_test)

print(f"MSE on train data:\t{mse(y_train, y_pred_train)}")
print(f"MSE on test data:\t{mse(y_test, y_pred_test)}")

%%time

rfr = RandomForestRegressor(n_estimators=500, max_depth=4, min_samples_split=5, min_samples_leaf=15, random_state=1)\
                            .fit(X_train, y_train)

y_pred_train = rfr.predict(X_train)
y_pred_test = rfr.predict(X_test)

print(f"MSE on train data:\t{mse(y_train, y_pred_train)}")
print(f"MSE on test data:\t{mse(y_test, y_pred_test)}")

### Predicting tags using Support Vector Regression

def get_prediction(X_train, y_train, X_test, model):
    
    model.fit(X_train, y_train)
    y_predicted = model.predict(X_test)
    
    return y_predicted

%%time

prediction, mse_train, mse_test = [], [], []

for i in range(len(y_flat.columns)):
    data_flat[f'target_{i}'] = y_flat[:][i]


train_df, test_df = train_test_split(data_flat, test_size=0.1, random_state=42)


for i in range(len(y_flat.columns)):
    y_train = train_df[f'target_{i}']
    X_train = train_df.iloc[:, :300]
    y_test = test_df[f'target_{i}']
    X_test = test_df.iloc[:, :300]
    
    svr = SVR(kernel='poly', degree=4, epsilon=0.05, C=3)
    
    y_pred = get_prediction(X_train, y_train, X_test, svr)

    prediction.append(y_pred)
    mse_train.append(mse(y_train, y_pred_train))
    mse_test.append(mse(y_test, y_pred_test))
    
    if len(prediction) % 30 == 0:
        print(f"Predictions done: {len(prediction)}")


y_full = pd.DataFrame(zip(*prediction))

print(f"Mean MSE on train data:\t{np.mean(mse_train)}")
print(f"Mean MSE on test data:\t{np.mean(mse_test)}")

y_test_pred, y_test_true = [], []

for i in range(y_full.shape[0]):
    y_test_pred.append(y_full.iloc[i, :].to_list())
    y_test_true.append(test_df.iloc[i, :300].to_list())

pred_words, true_words = [], []

for i in range(300):
    prediction = wv.most_similar(positive=[np.array(y_test_pred[i], dtype='float32')], restrict_vocab=100000, topn=1)[0][0]
    pred_words.append(prediction)
    true = tfidf_df.iloc[test_df.index[i]][1]
    true_words.append(true)

tags_comparison = pd.DataFrame(zip(pred_words, true_words), columns=['y_predicted', 'y_true'])

tags_comparison[41:60]

Possible further steps: train my own w2v model to make the vector space fit the corpus better; improve data processing and tokenization; predict tags based on several important words according to TF-IDF, not just one; use RNN.
